VAbenchmarks.py now runs with NEURON, NEST and PCSIM. Yay!
PCSIM does not yet support using a `NumpyRNG` instance to determine the connections, so the connectivity is not the same as in NEURON and NEST, but the spike statistics look convincingly similar.

Note that I have also added a `utility` module, which at the moment just contains a `notify()` function to send yourself an e-mail when the simulation has finished. Maybe this should go in NeuroTools?

git-svn-id: https://neuralensemble.org/svn/PyNN/trunk@61 1a7b9c28-352d-0410-8541-c879d6df8dcf