MAJOR CHANGES: merged multifile branch to trunk. The main changes:
(1) Simulator modules have been moved to a `src` subdirectory - this is to make distribution/installation of PyNN easier.

(2) Several of the modules have been split into multiple files, in their own subdirectories, e.g.: `nest2.py` --> `nest2/__init__.py`, `nest2/cells.py` and `nest2/connectors.py`. The reason for this is that the individual files were getting very long and difficult to navigate. It is not required, though. For example, I have kept `neuroml.py` as a single file because it was not too long. This change has broken a couple of the `neuron` unittests`, and I haven't run the `VAbenchmarks` test, but otherwise everything seems to work.

git-svn-id: https://neuralensemble.org/svn/PyNN/trunk@188 1a7b9c28-352d-0410-8541-c879d6df8dcf