Added a `dendritic_delay_fraction` argument (float between 0 and 1) to the `STDPMechanism` class. For point neurons and purely spike-based STDP rules, if we assume that the total delay between a spike occuring at the presynaptic cell soma/axon initial segment and the beginning of the EPSC at the postsynaptic soma is the sum of `d_axon`, `d_transmission` and `d_dendrite`, then `dendritic_delay_fraction=0` is equivalent to assuming that the back-propagating postsynaptic spike is felt instantaneously at the synapse (i.e. `d_dendrite` is zero) and `dendritic_delay_fraction=1` is equivalent to assuming that `d_axon` and `d_transmission` are both zero. In both cases we assume that forward transmission of the EPSP has the same delay as the back-propagating action potential, i.e. `d_dendrite==d_BPAP`.
Ideally, `dendritic_delay_fraction` should be able to vary on a per-connection basis (since `d_dendrite` should be more-or-less constant for a given cell class whereas `d_axon` is a function of the distance between the neurons), but until the underlying simulators support storing separate values for the axonal, transmission and dendritic delays, rather than a single 'synaptic' delay value, we make this value constant for all connections in a `Projection`.

Added a test to check that the `neuron` and `nest2` STDP implementations give comparable results.

git-svn-id: https://neuralensemble.org/svn/PyNN/trunk@248 1a7b9c28-352d-0410-8541-c879d6df8dcf