The attempted optimisation of setting parameter values in r408 broke the translation mechanism for those parameters whose values are computed from the values of more than one native parameter. It is for these "computed parameters" that we got all the native parameters.
I have fixed the bug while attempting to keep the optimisation by checking whether the list of parameters to be set contains any computed parameters, and only getting/translating all parameters in this case.

If the list contains only simple or scaled parameters, setting should be much faster.

Pierre, could you use your code profiling to check that this is the case?


git-svn-id: https://neuralensemble.org/svn/PyNN/trunk@410 1a7b9c28-352d-0410-8541-c879d6df8dcf