Fixed error ParameterSpace.columns()
Problem:
Setting attributes of shape `(n, 1)` fails for Projections.

Reason:
`lazyarray.partial_shape(...)` in `ParameterSpace.evaluate(...)` strips
dimensions containing only as single element. However, this information
is important for the `.columns()` and `.__iter__()` functions to
distinguish cases where there is a single row from a single column.

Solution:
The most straightforward solution was to copy `partial_shape(...)` from
lazyarray but to omit the dimension-reduction step. As the `.__iter__()`
and `.columns()` functions are the only ones accessing the
`_evaluated_shape` attribute.