Simplify interaction of random number generators and lazyarray
There are two changes here:
(1) previously, a random number generator with `parallel_safe=False` would always draw a reduced number of values when run with >1 MPI processes, according to the number of processes, unless the `mask_local` parameter was set to `False`. Now, a mask must be explicitly provided if you want to draw a reduced number of values (i.e. only those values consumed on that node), `mask` (renamed from `mask_local`) is either an array, or None, and can no longer take the value `False`.  
(2) moved the lazyarray behaviour for `RandomDistribution` from the lazyarray package to PyNN, to keep all the rng-related logic in one place.