Temporary fix for MPI bug in Projection.get
This is a rather subtle bug. The problem is that if gather is not 'all', on line 352 the method tries to address the array which is empty on nodes with rank != 0. Another solution would be to put that statement inside the preceding if-block and merge that if block with the enclosing one, so an empty 'values' is returned on nodes with rank != 0.